+++ package.json
@@ -22,7 +22,7 @@
     "react-native-reanimated": "^2.1.0",
     "react-native-screens": "^3.2.0",
     "react-native-vector-icons": "^8.1.0",
-    "react-native-web": "~0.13.12"
+    "react-native-web": "~0.13.12",
     "react-native-status-bar-height": "^2.6.0",
     "react-navigation": "^4.4.4",
     "react-navigation-stack": "react-navigation/stack"
