+++ src/screens/HomeScreen.js
@@ -1,10 +1,25 @@
-import * as React from 'react';
-import { Text, View } from 'react-native';
+import React from 'react'
+import Background from '../components/Background'
+import Logo from '../components/Logo'
+import Header from '../components/Header'
+import Paragraph from '../components/Paragraph'
+import Button from '../components/Button'
+import { logoutUser } from '../api/auth'
 
 export default function HomeScreen() {
-    return (
-      <View style = {{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-        <Text>Home!</Text>
-      </View>
-    );
+  return (
+//    <Background>
+//      <Logo />
+//      <Header>Let’s start</Header>
+//      <Paragraph>
+//        Main page
+//      </Paragraph>
+//      <Button mode="outlined" onPress={logoutUser}>
+//        Logout
+//      </Button>
+//    </Background>
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>Friends!</Text>
+    </View>
+  );
 }
